Moved old eotest tasks to archive folder (for now). Moved functions in crosstalkDMTask to crosstalk, and renamed crosstalkDMTask to crosstalkTask. Deleted unused utility functions in utils.
