+++ python/mixcoatl/crosstalkTask.py
@@ -1,458 +1,435 @@
-"""MixCOATL Tasks for crosstalk analysis.
+"""DM Crosstalk Tasks.
 
-TODO:
-    * Change logging implementation to match DM pipe task logging.
-    * Add docstrings and confirm compliance with LSP coding style guide.
-    * Update CrosstalkMatrix as needed.
+To Do: 
+    1. Determine method to restrict satellite crosstalk measurement to single side of sensor.
+    2. Clean up pipetask connection for rawExp.
 """
 import numpy as np
-from astropy.io import fits
-from scipy.ndimage.filters import gaussian_filter
-from sqlalchemy.orm.exc import NoResultFound
-import logging
-from datetime import datetime
-import os
+import copy
+from collections import defaultdict
 from skimage import feature
 from skimage.transform import hough_line, hough_line_peaks
 
-import lsst.pex.config as pexConfig
+from lsstDebug import getDebugFrame
+from lsst.afw.detection import FootprintSet, Threshold
+from lsst.afw.display import getDisplay
+from lsst.cp.pipe.utils import ddict2dict
+import lsst.afw.image as afwImage
+from lsst.ip.isr import CrosstalkCalib
+
 import lsst.pipe.base as pipeBase
-import lsst.eotest.image_utils as imutils
-from lsst.eotest.sensor.MaskedCCD import MaskedCCD
-from lsst.eotest.sensor.BrightPixels import BrightPixels
+import lsst.pipe.base.connectionTypes as cT
+from lsst.pex.config import Field, ListField, ConfigurableField
 
-from mixcoatl.crosstalk import CrosstalkMatrix, rectangular_mask, satellite_mask, crosstalk_fit
-from mixcoatl.utils import AMP2SEG, calculate_covariance
-from mixcoatl.database import Sensor, Segment, Result, db_session
+import mixcoatl.crosstalk as mixCrosstalk
 
-class InterCCDCrosstalkConfig(pexConfig.Config):
-    
-    length = pexConfig.Field("Length of postage stamps in y and x direction", int, default=200)
-    nsig = pexConfig.Field("Outlier rejection sigma threshold", float, default=5.0)
-    num_iter = pexConfig.Field("Number of least square iterations", int, default=3)
-    threshold = pexConfig.Field("Aggressor spot mean signal threshold", float, default=40000.)
-    verbose = pexConfig.Field("Turn verbosity on", bool, default=True)
-
-class InterCCDCrosstalkTask(pipeBase.Task):
-
-    ConfigClass = InterCCDCrosstalkConfig
-    _DefaultName = "InterCCDCrosstalkTask"
-
-    def run(self, sensor_id1, infiles1, gains1, bias_frame1=None, 
-            dark_frame1=None, crosstalk_matrix_file=None, **kwargs):
-
-        ## Parse kwargs for separate victim CCD inputs
-        try:
-            sensor_id2 = kwargs['sensor_id2']
-        except KeyError:
-            sensor_id2 = sensor_id1
-            infiles2 = infiles1
-            gains2 = gains1
-            bias_frame2 = bias_frame1
-            dark_frame2 = dark_frame1
-        else:
-            infiles2 = kwargs['infiles2']
-            gains2 = kwargs['gains2']
-            bias_frame2 = kwargs['bias_frame2']
-            dark_frame2 = kwargs['dark_frame2']
-
-        infiles_list = [(infiles1[i], infiles2[i]) for i in range(len(infiles1))]
-        
-        all_amps = imutils.allAmps(infiles1[0])
-
-        ## Create new matrix or modify existing
-        if crosstalk_matrix_file is not None:
-            crosstalk_matrix = CrosstalkMatrix.from_fits(crosstalk_matrix_file)
-            outfile = crosstalk_matrix_file
-        else:
-            crosstalk_matrix = CrosstalkMatrix(sensor_id1, 
-                                               victim_id=sensor_id2,
-                                               namps=len(all_amps))
-            outfile = self.config.outfile
-
-        for infile1, infile2 in infiles_list:
-            ccd1 = MaskedCCD(infile1, bias_frame=bias_frame1, 
-                             dark_frame=dark_frame1)
-            ccd2 = MaskedCCD(infile2, bias_frame=bias_frame2, 
-                             dark_frame=dark_frame2)      
-            num_aggressors = 0
-            signals = []
-
-            ## Search each amp for aggressor
-            for i in all_amps:
-                imarr1 = ccd1.unbiased_and_trimmed_image(i).getImage().getArray()*gains1[i]
-                smoothed = gaussian_filter(imarr1, 20)
-                y, x = np.unravel_index(smoothed.argmax(), smoothed.shape)
-                stamp1 = make_stamp(imarr1, y, x)
-                ly, lx = stamp1.shape
-                Y, X = np.ogrid[-ly/2:ly/2, -lx/2:lx/2]
-                mask = X*X + Y*Y <= 20*20
-                signal = np.mean(stamp1[mask])
-
-                if signal > self.config.threshold:
-                    signals.append(signal)
-                    row = {}
-
-                    ## Calculate crosstalk for each victim amp
-                    for j in all_amps:
-                        imarr2 = ccd2.unbiased_and_trimmed_image(j).getImage().getArray()*gains2[j]
-
-                        stamp2 = make_stamp(imarr2, y, x)
-                        row[j] = crosstalk_fit(stamp1, stamp2, noise=7.0,
-                                                     num_iter=self.config.num_iter,
-                                                     nsig=self.config.nsig)
-
-                    crosstalk_matrix.set_row(i, row)
-                    if num_aggressors == self.config.aggressors_per_image: 
-                        break
-
-        crosstalk_matrix.signal = np.median(np.asarray(signals))
-        if sensor_id1 == sensor_id2:
-            crosstalk_matrix.set_diagonal(0.)
-        crosstalk_matrix.write_fits(outfile, overwrite=True)
-
-class CrosstalkSpotConfig(pipeBase.PipelineTaskConfig,
-                          pipelineConnections=pipeBase.PipelineTaskConnections):
-
-    database = pexConfig.Field(
-        dtype=str,
-        doc="SQL database DB file.", 
-        default='crosstalk.db'
+class CrosstalkColumnConnections(pipeBase.PipelineTaskConnections,
+                                 dimensions=("instrument", "exposure", "detector")):
+    inputExp = cT.Input(
+        name="crosstalkInputs",
+        doc="Input post-ISR processed exposure to measure crosstalk from.",
+        storageClass="Exposure",
+        dimensions=("instrument", "exposure", "detector"),
+        multiple=False,
     )
-    maskLengthX = pexConfig.Field(
-        dtype=int,
-        doc="Length of rectangular mask in x-direction.",
-        default=200
+    rawExp = cT.Input(
+        name="rawInputs",
+        doc="Input raw exposure to measure noise covariance from.",
+        storageClass="Exposure",
+        dimensions=("instrument", "exposure", "detector"),
+        multiple=False,
     )
-    maskLengthY = pexConfig.Field(
-        dtype=int,
-        doc="Length of rectangular mask in y-direction.", 
-        default=200
+    outputRatios = cT.Output(
+        name="crosstalkRatios",
+        doc="Extracted crosstalk pixel ratios.",
+        storageClass="StructuredDataDict",
+        dimensions=("instrument", "exposure", "detector"),
     )
-    threshold = pexConfig.Field(
-        dtype=float,
-        doc="Aggressor spot mean signal threshold.", 
-        default=50000.
+    outputFluxes = cT.Output(
+        name="crosstalkFluxes",
+        doc="Source pixel fluxes used in ratios.",
+        storageClass="StructuredDataDict",
+        dimensions=("instrument", "exposure", "detector"),
     )
-    covarianceCorrectionType = pexConfig.ChoiceField(
-        dtype=str,
-        doc="The method for correcting correlated read noise.",
-        default='NONE',
-        allowed={
-            "NONE": "No correction performed.",
-            "RANDOM": "Correct using a random seed.",
-            "SEED" : "Correct using a fixed seed."
-        }
+    outputZOffsets = cT.Output(
+        name="crosstalkBackgroundZOffsets",
+        doc="Z offset parameters used in background model.",
+        storageClass="StructuredDataDict",
+        dimensions=("instrument", "exposure", "detector"),
+    )
+    outputYTilts = cT.Output(
+        name="crosstalkBackgroundYTilts",
+        doc="Y tilt parameters used in background model.",
+        storageClass="StructuredDataDict",
+        dimensions=("instrument", "exposure", "detector"),
+    )
+    outputXTilts = cT.Output(
+        name="crosstalkBackgroundXTilts",
+        doc="X tilt parameters used in background model.",
+        storageClass="StructuredDataDict",
+        dimensions=("instrument", "exposure", "detector"),
     )
 
-class CrosstalkSpotTask(pipeBase.Task):
+    def __init__(self, *, config=None):
+        super().__init__(config=config)
 
-    ConfigClass = CrosstalkSpotConfig
-    _DefaultName = "CrosstalkSpotTask"
-
-    def run(self, sensor_name, infile, bias_frame=None, dark_frame=None, linearity_correction=None):
-
-        if self.config.covarianceCorrectionType == 'NONE':
-            correct_covariance = False
-            seed = None
-        elif self.config.covarianceCorrectionType == 'RANDOM':
-            correct_covariance = True
-            seed = None
-        elif self.config.covarianceCorrectionType == 'SEED':
-            correct_covariance = True
-            seed = 189
-
-        ## Get sensor information from header
-        all_amps = imutils.allAmps(infile)
-        with fits.open(infile) as hdulist:
-            lsst_num = hdulist[0].header['LSST_NUM']
-            teststand = hdulist[0].header['TSTAND']
-            manufacturer = lsst_num[:3]
-
-        ## Interface with SQL database
-        database = self.config.database
-        with db_session(database) as session:
-
-            ## Get sensor from database
-            try:
-                sensor = Sensor.from_db(session, sensor_name=sensor_name)
-            except NoResultFound:
-                sensor = Sensor(sensor_name=sensor_name, lsst_num=lsst_num, manufacturer=manufacturer, 
-                                namps=len(all_amps))
-                sensor.segments = {i : Segment(segment_name=AMP2SEG[i], amplifier_number=i) for i in all_amps}
-                sensor.add_to_db(session)
-                session.commit()
-
-            ## Set configuration and analysis settings
-            ccd = MaskedCCD(infile, bias_frame=bias_frame, dark_frame=dark_frame, 
-                            linearity_correction=linearity_correction)
-
-            ## Aggressor amplifiers
-            for i in all_amps:
-
-                aggressor_imarr = ccd.unbiased_and_trimmed_image(i).getImage().getArray()
-
-                ## Find aggressor regions
-                smoothed = gaussian_filter(aggressor_imarr, 20)
-                y, x = np.unravel_index(smoothed.argmax(), smoothed.shape)
-                select = rectangular_mask(aggressor_imarr, y, x, ly=self.config.maskLengthY, 
-                                        lx=self.config.maskLengthX)
-                signal = np.max(smoothed)
-                if signal < self.config.threshold:
-                    continue
-    
-                ## Victim amplifiers
-                for j in all_amps:
-
-                    covariance = calculate_covariance(ccd, i, j)
-                    victim_imarr = ccd.unbiased_and_trimmed_image(j).getImage().getArray()
-
-                    ## Add crosstalk result to database
-                    res = crosstalk_fit(aggressor_imarr, victim_imarr, select, covariance=covariance,
-                                        correct_covariance=correct_covariance, seed=seed)
-                    result = Result(aggressor_id=sensor.segments[i].id, victim_id=sensor.segments[j].id,
-                                    aggressor_signal=signal, coefficient=res[0], error=res[4],
-                                    methodology='MODEL_LSQ', teststand=teststand, image_type='spot',
-                                    analysis='CrosstalkSpotTask', image_file=os.path.basename(infile),
-                                    z_offset=res[1], y_tilt=res[2], x_tilt=res[3])
-                    result.add_to_db(session)
-                    
 class CrosstalkColumnConfig(pipeBase.PipelineTaskConfig,
-                            pipelineConnections=pipeBase.PipelineTaskConnections):
-
-    database = pexConfig.Field(
-        dtype=str,
-        doc="SQL database DB file.", 
-        default='crosstalk.db'
+                            pipelineConnections=CrosstalkColumnConnections):
+    """Configuration for the measurement of pixel ratios.
+    """
+    correctNoiseCovariance = Field(
+        dtype=bool,
+        default=False,
+        doc="Correct the effect of correlated read noise."
     )
-    maskLengthX = pexConfig.Field(
-        dtype=int,
-        doc="Length of rectangular mask in x-direction.",
-        default=100
-    )
-    maskLengthY = pexConfig.Field(
-        dtype=int,
-        doc="Length of postage stamp mask in y-direction.", 
-        default=2000
-    )
-    threshold = pexConfig.Field(
+    threshold = Field(
         dtype=float,
-        doc="Aggressor column mean signal threshold.",
-        default=1000.
+        default=30000,
+        doc="Minimum level of source pixels for which to measure crosstalk."
     )
-    covarianceCorrectionType = pexConfig.ChoiceField(
+    maskLengthX = Field(
+        dtype=int,
+        default=100,
+        doc="Length of rectangular mask in x-direction."
+    )
+    maskLengthY = Field(
+        dtype=int,
+        default=2000,
+        doc="Length of postage stamp mask in y-direction."
+    )
+    ignoreSaturatedPixels = Field(
+        dtype=bool,
+        default=True,
+        doc="Should saturated pixels be ignored?"
+    )
+    badMask = ListField(
         dtype=str,
-        doc="The method for correcting correlated read noise.",
-        default='NONE',
-        allowed={
-            "NONE": "No correction performed.",
-            "RANDOM": "Correct using a random seed.",
-            "SEED" : "Correct using a fixed seed."
-        }
+        default=["BAD", "INTRP"],
+        doc="Mask planes to ignore when identifying source pixels."
+    )
+    isTrimmed = Field(
+        dtype=bool,
+        default=True,
+        doc="Is the input exposure trimmed?"
     )
 
-class CrosstalkColumnTask(pipeBase.PipelineTask):
+    def validate(self):
+        super().validate()
 
+        # Ensure the handling of the SAT mask plane is consistent
+        # with the ignoreSaturatedPixels value.
+        if self.ignoreSaturatedPixels:
+            if 'SAT' not in self.badMask:
+                self.badMask.append('SAT')
+        else:
+            if 'SAT' in self.badMask:
+                self.badMask = [mask for mask in self.badMask if mask != 'SAT']
+
+class CrosstalkColumnTask(pipeBase.PipelineTask,
+                          pipeBase.CmdLineTask):
+    """Task to measure pixel ratios to find crosstalk.
+    """
     ConfigClass = CrosstalkColumnConfig
-    _DefaultName = "CrosstalkColumnTask"
+    _DefaultName = 'cpCrosstalkColumn'
 
-    def run(self, sensor_name, infile, bias_frame=None, dark_frame=None, linearity_correction=None):
+    @pipeBase.timeMethod
+    def run(self, inputExp, rawExp):
 
-        if self.config.covarianceCorrectionType == 'NONE':
-            correct_covariance = False
-            seed = None
-        elif self.config.covarianceCorrectionType == 'RANDOM':
-            correct_covariance = True
-            seed = None
-        elif self.config.covarianceCorrectionType == 'SEED':
-            correct_covariance = True
-            seed = 189
+        ## run() method
+        outputRatios = defaultdict(lambda: defaultdict(dict))
+        outputFluxes = defaultdict(lambda: defaultdict(dict))
+        outputZOffsets = defaultdict(lambda: defaultdict(dict))
+        outputYTilts = defaultdict(lambda: defaultdict(dict))
+        outputXTilts = defaultdict(lambda: defaultdict(dict))
 
-        ## Get sensor information from header
-        all_amps = imutils.allAmps(infile)
-        with fits.open(infile) as hdulist:
-            lsst_num = hdulist[0].header['LSST_NUM']
-            teststand = hdulist[0].header['TSTAND']
-            manufacturer = lsst_num[:3]
+        badPixels = list(self.config.badMask)
 
-        ## Interface with SQL database
-        database = self.config.database
-        with db_session(database) as session:
+        targetDetector = inputExp.getDetector()
+        targetChip = targetDetector.getName()
+        targetIm = inputExp.getMaskedImage()
 
-            ## Get sensor from database
-            try:
-                sensor = Sensor.from_db(session, sensor_name=sensor_name)
-            except NoResultFound:
-                sensor = Sensor(sensor_name=sensor_name, lsst_num=lsst_num, manufacturer=manufacturer,
-                                namps=len(all_amps))
-                sensor.segments = {i : Segment(segment_name=AMP2SEG[i], amplifier_number=i) for i in all_amps}
-                sensor.add_to_db(session)
-                session.commit()
+        ## loop on sourceExp would go here
+        sourceExp = inputExp ## Ignore other exposures for now
+        sourceDetector = sourceExp.getDetector()
+        sourceChip = sourceDetector.getName()
+        sourceIm = sourceExp.getMaskedImage()
+        bad = sourceIm.getMask().getPlaneBitMask(badPixels)
+        self.log.info("Measuring crosstalk from source: %s", sourceChip)
 
-            ccd = MaskedCCD(infile, bias_frame=bias_frame, dark_frame=dark_frame, 
-                            linearity_correction=linearity_correction)
+        ratioDict = defaultdict(lambda: defaultdict(list))
+        zoffsetDict = defaultdict(lambda: defaultdict(list))
+        ytiltDict = defaultdict(lambda: defaultdict(list))
+        xtiltDict = defaultdict(lambda: defaultdict(list))
 
-            ## Aggressor amplifiers
-            for i in all_amps:
+        extractedCount = 0
 
-                ## Find aggressor regions
-                exptime = 1
-                gain = 1
-                bp = BrightPixels(ccd, i, exptime, gain, ethresh=self.config.threshold)
-                pixels, columns = bp.find()
-                if len(columns) == 0:
+        for sourceAmp in sourceDetector:
+            sourceAmpName = sourceAmp.getName()
+            sourceAmpImage = sourceIm[sourceAmp.getBBox()]
+            sourceMask = sourceAmpImage.mask.array
+            sourceAmpArray = sourceAmpImage.image.array
+
+            columns = mixCrosstalk.find_bright_columns(sourceAmpArray, self.config.threshold)
+            if len(columns) == 0: continue
+            select = mixCrosstalk.rectangular_mask(sourceAmpArray, 1000, columns[0],
+                                                   ly=self.config.maskLengthY, lx=self.config.maskLengthX)
+            signal = np.mean(sourceAmpArray[:, columns[0]])
+            self.log.debug("  Source amplifier: %s", sourceAmpName)
+
+            outputFluxes[sourceChip][sourceAmpName] = [float(signal)]
+
+            for targetAmp in targetDetector:
+                # iterate over targetExposure
+                targetAmpName = targetAmp.getName()
+                if sourceAmpName == targetAmpName and sourceChip == targetChip:
+                    ratioDict[sourceAmpName][targetAmpName] = []
                     continue
-                col = columns[0]
+                self.log.debug("    Target amplifier: %s", targetAmpName)
 
-                aggressor_imarr = ccd.unbiased_and_trimmed_image(i).getImage().getArray()
-                signal = np.mean(aggressor_imarr[:, col])
-                mask = rectangular_mask(aggressor_imarr, 1000, col, ly=self.config.maskLengthY,
-                                        lx=self.config.maskLengthX)
-                
-                ## Victim amplifiers
-                for j in all_amps:
+                if self.config.correctNoiseCovariance:
+                    covariance = mixCrosstalk.calculate_covariance(rawExp, sourceAmp, targetAmp)
+                else:
+                    covariance = np.asarray([[7.0, 0.0], [0.0, 7.0]])
 
-                    covariance = calculate_covariance(ccd, i, j)
-                    victim_imarr = ccd.unbiased_and_trimmed_image(j).getImage().getArray()
-                    
-                    ## Add crosstalk result to database
-                    res = crosstalk_fit(aggressor_imarr, victim_imarr, mask, covariance=covariance,
-                                        correct_covariance=correct_covariance, seed=seed)
-                    result = Result(aggressor_id=sensor.segments[i].id, victim_id=sensor.segments[j].id,
-                                    aggressor_signal=signal, coefficient=res[0], error=res[4], 
-                                    methodology='MODEL_LSQ', image_type='brightcolumn',
-                                    teststand=teststand, analysis='CrosstalkColumnTask', 
-                                    image_file=os.path.basename(infile), z_offset=res[1], y_tilt=res[2], 
-                                    x_tilt=res[3])
-                    result.add_to_db(session)
+                targetAmpImage = CrosstalkCalib.extractAmp(targetIm.image,
+                                                           targetAmp, sourceAmp,
+                                                           isTrimmed=self.config.isTrimmed)
+                targetAmpArray = targetAmpImage.array
+                results = mixCrosstalk.crosstalk_fit(sourceAmpArray, targetAmpArray, select, 
+                                                     covariance=covariance,
+                                                     correct_covariance=self.config.correctNoiseCovariance, 
+                                                     seed=189)
+
+                ratioDict[targetAmpName][sourceAmpName] = [float(results[0])]
+                zoffsetDict[targetAmpName][sourceAmpName] = [float(results[1])]
+                ytiltDict[targetAmpName][sourceAmpName] = [float(results[2])]
+                xtiltDict[targetAmpName][sourceAmpName] = [float(results[3])]
+                extractedCount += 1
+
+        self.log.info("Extracted %d pixels from %s -> %s",
+                      extractedCount, sourceChip, targetChip)
+        outputRatios[targetChip][sourceChip] = ratioDict
+        outputZOffsets[targetChip][sourceChip] = zoffsetDict
+        outputYTilts[targetChip][sourceChip] = ytiltDict
+        outputXTilts[targetChip][sourceChip] = xtiltDict
+        
+        return pipeBase.Struct(
+            outputRatios=ddict2dict(outputRatios),
+            outputFluxes=ddict2dict(outputFluxes),
+            outputZOffsets=ddict2dict(outputZOffsets),
+            outputYTilts=ddict2dict(outputYTilts),
+            outputXTilts=ddict2dict(outputXTilts)
+        )
+
+class CrosstalkSatelliteConnections(pipeBase.PipelineTaskConnections,
+                                    dimensions=("instrument", "exposure", "detector")):
+    inputExp = cT.Input(
+        name="crosstalkInputs",
+        doc="Input post-ISR processed exposure to measure crosstalk from.",
+        storageClass="Exposure",
+        dimensions=("instrument", "exposure", "detector"),
+        multiple=False,
+    )
+    rawExp = cT.Input(
+        name="rawInputs",
+        doc="Input raw exposure to measure noise covariance from.",
+        storageClass="Exposure",
+        dimensions=("instrument", "exposure", "detector"),
+        multiple=False,
+    )
+    outputRatios = cT.Output(
+        name="crosstalkRatios",
+        doc="Extracted crosstalk pixel ratios.",
+        storageClass="StructuredDataDict",
+        dimensions=("instrument", "exposure", "detector"),
+    )
+    outputFluxes = cT.Output(
+        name="crosstalkFluxes",
+        doc="Source pixel fluxes used in ratios.",
+        storageClass="StructuredDataDict",
+        dimensions=("instrument", "exposure", "detector"),
+    )
+    outputZOffsets = cT.Output(
+        name="crosstalkBackgroundZOffsets",
+        doc="Z offset parameters used in background model.",
+        storageClass="StructuredDataDict",
+        dimensions=("instrument", "exposure", "detector"),
+    )
+    outputYTilts = cT.Output(
+        name="crosstalkBackgroundYTilts",
+        doc="Y tilt parameters used in background model.",
+        storageClass="StructuredDataDict",
+        dimensions=("instrument", "exposure", "detector"),
+    )
+    outputXTilts = cT.Output(
+        name="crosstalkBackgroundXTilts",
+        doc="X tilt parameters used in background model.",
+        storageClass="StructuredDataDict",
+        dimensions=("instrument", "exposure", "detector"),
+    )
+
+    def __init__(self, *, config=None):
+        super().__init__(config=config)
 
 class CrosstalkSatelliteConfig(pipeBase.PipelineTaskConfig,
-                               pipelineConnections=pipeBase.PipelineTaskConnections):
-
-    database = pexConfig.Field(
-        dtype=str,
-        doc="SQL database DB file.", 
-        default='crosstalk.db'
+                               pipelineConnections=CrosstalkSatelliteConnections):
+    """Configuration for the measurement of pixel ratios.
+    """
+    correctNoiseCovariance = Field(
+        dtype=bool,
+        default=False,
+        doc="Correct the effect of correlated read noise."
     )
-    maskWidth = pexConfig.Field(
+    thresholdLow = Field(
+        dtype=float,
+        default=1,
+        doc="Low threshold for Canny edge detection."
+    )
+    thresholdHigh = Field(
+        dtype=float,
+        default=15,
+        doc="High threshold for Canny edge detection."
+    )
+    maskWidth = Field(
         dtype=int,
-        doc="Width of satellite streak mask.", 
-        default=50
+        default=50,
+        doc="One-sided width of satellite streak mask."
     )
-    cannySigma = pexConfig.Field(
+    cannySigma = Field(
         dtype=float,
         doc="Gaussian smoothing sigma for Canny edge detection.", 
         default=15.
     )
-    thresholdLow = pexConfig.Field(
-        dtype=float,
-        doc="Low threshold for Canny edge detection.", 
-        default=1
-    )
-    thresholdHigh = pexConfig.Field(
-        dtype=float,
-        doc="High threshold for Canny edge detection.", 
-        default=15
-    )
-    restrictSide = pexConfig.Field(
+    ignoreSaturatedPixels = Field(
         dtype=bool,
-        doc="Restrict crosstalk measurement to segment pairs on a single side", 
-        default=True
+        default=True,
+        doc="Should saturated pixels be ignored?"
     )
-    covarianceCorrectionType = pexConfig.ChoiceField(
+    badMask = ListField(
         dtype=str,
-        doc="The method for correcting correlated read noise.", 
-        default='NONE',
-        allowed={
-            "NONE": "No correction performed.",
-            "RANDOM": "Correct using a random seed.",
-            "SEED" : "Correct using a fixed seed."
-        }
+        default=["BAD", "INTRP"],
+        doc="Mask planes to ignore when identifying source pixels."
+    )
+    isTrimmed = Field(
+        dtype=bool,
+        default=True,
+        doc="Is the input exposure trimmed?"
     )
 
-class CrosstalkSatelliteTask(pipeBase.Task):
+    def validate(self):
+        super().validate()
 
+        # Ensure the handling of the SAT mask plane is consistent
+        # with the ignoreSaturatedPixels value.
+        if self.ignoreSaturatedPixels:
+            if 'SAT' not in self.badMask:
+                self.badMask.append('SAT')
+        else:
+            if 'SAT' in self.badMask:
+                self.badMask = [mask for mask in self.badMask if mask != 'SAT']
+
+class CrosstalkSatelliteTask(pipeBase.PipelineTask,
+                             pipeBase.CmdLineTask):
+    """Task to measure pixel ratios to find crosstalk.
+    """
     ConfigClass = CrosstalkSatelliteConfig
-    _DefaultName = "CrosstalkSatelliteTask"
+    _DefaultName = 'cpCrosstalkSatellite'
 
-    def run(self, sensor_name, infile, bias_frame=None, dark_frame=None, linearity_correction=None):
+    @pipeBase.timeMethod
+    def run(self, inputExp, rawExp):
 
-        if self.config.covarianceCorrectionType == 'NONE':
-            correct_covariance = False
-            seed = None
-        elif self.config.covarianceCorrectionType == 'RANDOM':
-            correct_covariance = True
-            seed = None
-        elif self.config.covarianceCorrectionType == 'SEED':
-            correct_covariance = True
-            seed = 189
+        ## run() method
+        outputRatios = defaultdict(lambda: defaultdict(dict))
+        outputFluxes = defaultdict(lambda: defaultdict(dict))
+        outputZOffsets = defaultdict(lambda: defaultdict(dict))
+        outputYTilts = defaultdict(lambda: defaultdict(dict))
+        outputXTilts = defaultdict(lambda: defaultdict(dict))
 
-        ## Get sensor information from header
-        all_amps = imutils.allAmps(infile)
-        with fits.open(infile) as hdulist:
-            lsst_num = hdulist[0].header['LSST_NUM']
-            teststand = hdulist[0].header['TSTAND']
-            manufacturer = lsst_num[:3]
+        badPixels = list(self.config.badMask)
 
-        ## Interface with SQL database
-        database = self.config.database
-        with db_session(database) as session:
+        targetDetector = inputExp.getDetector()
+        targetChip = targetDetector.getName()
+        targetIm = inputExp.getMaskedImage()
 
-            ## Get sensor from database
-            try:
-                sensor = Sensor.from_db(session, sensor_name=sensor_name)
-            except NoResultFound:
-                sensor = Sensor(sensor_name=sensor_name, lsst_num=lsst_num, manufacturer=manufacturer,
-                                namps=len(all_amps))
-                sensor.segments = {i : Segment(segment_name=AMP2SEG[i], amplifier_number=i) for i in all_amps}
-                sensor.add_to_db(session)
-                session.commit()
+        ## loop on sourceExp would go here
+        sourceExp = inputExp ## Ignore other exposures for now
+        sourceDetector = sourceExp.getDetector()
+        sourceChip = sourceDetector.getName()
+        sourceIm = sourceExp.getMaskedImage()
+        bad = sourceIm.getMask().getPlaneBitMask(badPixels)
+        self.log.info("Measuring crosstalk from source: %s", sourceChip)
 
-            ccd = MaskedCCD(infile, bias_frame=bias_frame, dark_frame=dark_frame, 
-                            linearity_correction=linearity_correction)
+        ratioDict = defaultdict(lambda: defaultdict(list))
+        zoffsetDict = defaultdict(lambda: defaultdict(list))
+        ytiltDict = defaultdict(lambda: defaultdict(list))
+        xtiltDict = defaultdict(lambda: defaultdict(list))
 
-            ## Aggressor amplifiers
-            for i in all_amps:
+        extractedCount = 0
 
-                aggressor_imarr = ccd.unbiased_and_trimmed_image(i).getImage().getArray()
+        for sourceAmp in sourceDetector:
+            sourceAmpName = sourceAmp.getName()
+            sourceAmpImage = sourceIm[sourceAmp.getBBox()]
+            sourceMask = sourceAmpImage.mask.array
+            sourceAmpArray = sourceAmpImage.image.array
 
-                ## Find aggressor regions
-                tested_angles = np.linspace(-np.pi / 2, np.pi / 2, 1000)
-                edges = feature.canny(aggressor_imarr, sigma=self.config.cannySigma, 
-                                      low_threshold=self.config.thresholdLow, 
-                                      high_threshold=self.config.thresholdHigh)
-                h, theta, d = hough_line(edges, theta=tested_angles)
-                _, angle, dist = hough_line_peaks(h, theta, d)
+            tested_angles = np.linspace(-np.pi / 2, np.pi / 2, 1000)
+            edges = feature.canny(sourceAmpArray, sigma=self.config.cannySigma, 
+                                  low_threshold=self.config.thresholdLow, 
+                                  high_threshold=self.config.thresholdHigh)
+            h, theta, d = hough_line(edges, theta=tested_angles)
+            _, angle, dist = hough_line_peaks(h, theta, d)
 
-                if len(angle) != 2:
+            if len(angle) != 2:
+                continue
+
+            mean_angle = np.mean(angle)
+            mean_dist = np.mean(dist)
+            select = mixCrosstalk.satellite_mask(sourceAmpArray, mean_angle, mean_dist, 
+                                                 width=self.config.maskWidth)
+            signal = np.max(sourceAmpArray[select])
+            self.log.debug("  Source amplifier: %s", sourceAmpName)
+
+            outputFluxes[sourceChip][sourceAmpName] = [float(signal)]
+
+            for targetAmp in targetDetector:
+                # iterate over targetExposure
+                targetAmpName = targetAmp.getName()
+                if sourceAmpName == targetAmpName and sourceChip == targetChip:
+                    ratioDict[sourceAmpName][targetAmpName] = []
                     continue
+                self.log.debug("    Target amplifier: %s", targetAmpName)
 
-                mean_angle = np.mean(angle)
-                mean_dist = np.mean(dist)
-                select = satellite_mask(aggressor_imarr, mean_angle, mean_dist, width=self.config.maskWidth)
-                signal = np.max(aggressor_imarr[select])
-                
-                ## Victim amplifiers
-                if self.config.restrictSide:
-                    if i < 9:
-                        vic_amps = range(1, 9)
-                    else:
-                        vic_amps = range(9, 17)
+                if self.config.correctNoiseCovariance:
+                    covariance = mixCrosstalk.calculate_covariance(rawExp, sourceAmp, targetAmp)
                 else:
-                    vic_amps = all_amps
+                    covariance = np.asarray([[7.0, 0.0], [0.0, 7.0]])
 
-                for j in vic_amps:
+                targetAmpImage = CrosstalkCalib.extractAmp(targetIm.image,
+                                                           targetAmp, sourceAmp,
+                                                           isTrimmed=self.config.isTrimmed)
+                targetAmpArray = targetAmpImage.array
+                results = mixCrosstalk.crosstalk_fit(sourceAmpArray, targetAmpArray, select, 
+                                                     covariance=covariance, 
+                                                     correct_covariance=self.config.correctNoiseCovariance, 
+                                                     seed=189)
 
-                    covariance = calculate_covariance(ccd, i, j)
-                    victim_imarr = ccd.unbiased_and_trimmed_image(j).getImage().getArray()
+                ratioDict[targetAmpName][sourceAmpName] = [float(results[0])]
+                zoffsetDict[targetAmpName][sourceAmpName] = [float(results[1])]
+                ytiltDict[targetAmpName][sourceAmpName] = [float(results[2])]
+                xtiltDict[targetAmpName][sourceAmpName] = [float(results[3])]
+                extractedCount += 1
 
-                    res = crosstalk_fit(aggressor_imarr, victim_imarr, select, covariance=covariance,
-                                        correct_covariance=correct_covariance, seed=seed)
-
-                    ## Add result to database
-                    result = Result(aggressor_id=sensor.segments[i].id, victim_id=sensor.segments[j].id,
-                                    aggressor_signal=signal, coefficient=res[0], error=res[4], 
-                                    methodology='MODEL_LSQ', teststand=teststand, image_type='satellite',
-                                    analysis='CrosstalkSatelliteTask', image_file=os.path.basename(infile), 
-                                    z_offset=res[1], y_tilt=res[2], x_tilt=res[3])
-                    result.add_to_db(session)
+        self.log.info("Extracted %d pixels from %s -> %s",
+                      extractedCount, sourceChip, targetChip)
+        outputRatios[targetChip][sourceChip] = ratioDict
+        outputZOffsets[targetChip][sourceChip] = zoffsetDict
+        outputYTilts[targetChip][sourceChip] = ytiltDict
+        outputXTilts[targetChip][sourceChip] = xtiltDict
+        
+        return pipeBase.Struct(
+            outputRatios=ddict2dict(outputRatios),
+            outputFluxes=ddict2dict(outputFluxes),
+            outputZOffsets=ddict2dict(outputZOffsets),
+            outputYTilts=ddict2dict(outputYTilts),
+            outputXTilts=ddict2dict(outputXTilts)
+        )
